Fix Z16F context structure
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@556 42af7a65-404d-4744-a932-0658087f49c3

+++ arch/z16/include/types.h
@@ -69,7 +69,7 @@ typedef unsigned int uint32;
  * irqsave()
  */
 
-typedef unsigned int irqstate_t;
+typedef unsigned short irqstate_t;
 
 #endif /* __ASSEMBLY__ */
